fix(#38): 페이지들의 Client/Server 이중 평가 문제 해결
