+++ libs/ui/src/lib/pages/MainRoomClient.tsx
@@ -1,0 +1,23 @@
+'use client';
+
+import { MainRoomPage } from './MainRoomPage';
+import { projectSidebarList } from '../../context';
+import { useParams } from 'next/navigation';
+
+export function MainRoomClient() {
+  const params = useParams();
+  const serverId = Number(params.serverId);
+
+  const roomNameFind = Number.isNaN(serverId)
+    ? null
+    : projectSidebarList.find((room) => room.team_id === serverId)?.team_name ??
+      null;
+
+  return (
+    <>
+      <div className="flex-1 flex overflow-hidden">
+        {roomNameFind && <MainRoomPage />}
+      </div>
+    </>
+  );
+}
