+++ libs/ui/src/lib/pages/index.ts
@@ -4,3 +4,5 @@ export * from './SummaryPage';
 export * from './ChatRoomPage';
 export * from './ServerProfilePage';
 export * from './LoginPage';
+export * from './MainRoomClient';
+export * from './SignUpPage';
